Use consistent date format, improve date bounds
For JavaScript Date() function, YYYY-MM-DD implies a UTC date whereas MM/DD/YYYY implies a local date. This commit parses the strings and creates the Date object with Date(Year, Month, Day) which creates a local date.

Also, this commit changes the filter function start date and end date bounds to be at the beginning and end of the bounds. This allows for searching for a single day. Previously, the bounds were both at the beginning of the day, so you had to know to add one day to the end in order to search for a single day, i.e. search July 10 - July 11 in order to search for July 10.